backend/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
sentence-transformers
numpy
requests
google-generativeai
python-dotenv
